router: create module
